Comentar alternativas de cálculo de suma de elementos de una lista de objetos en el método obtenerNumeroVentasIngresosDiarios.

+++ src/main/java/com/codoacodo/flysky/demo/dto/response/VueloDto.java
@@ -16,7 +16,7 @@ public class VueloDto {
 
     private Integer numeroVuelo;
     private Boolean disponible;
-    private Integer capacidad;
+    private Long capacidad;
     private String aerolinea;
     private LocalDateTime fechaHoraPartida;
     private LocalDateTime fechaHoraLlegada;
